Update: 컬러 선택 기능 추가
