+++ Frontend/src/pages/category/type.ts
@@ -1,0 +1,14 @@
+export interface Category {
+  category: string;
+  color: string;
+}
+
+export interface CategoryHeaderPropsType {
+  color: string;
+  categoryArr: Category[];
+  handleClose: () => void;
+  open: boolean;
+  setColor: React.Dispatch<React.SetStateAction<string>>;
+  setCategoryArr: React.Dispatch<React.SetStateAction<Category[]>>;
+  setOpen: React.Dispatch<React.SetStateAction<boolean>>;
+}
